+++ src/views/TutorialsListView.vue
@@ -10,6 +10,19 @@
                 </div>
             </div>
         </div>
+        <div class="col-md-12">
+            <div class="mb-3" align="center">
+                <h6 class="text-center">Items per Page:</h6>
+                <select v-model="pageSize" @change="handlePageSizeChange($event)">
+                    <option v-for="size in pageSize" :key="size" :value="size">
+                        {{ size }}
+                    </option>
+                </select>
+            </div>
+            <b-pagination v-model="page" :total-rows="count" :per-page="pageSize" prev-text="Prev" next-text="Next"
+                @change="handlePageChange" align="center"></b-pagination>
+        </div>
+
         <div class="col-md-6">
             <h4>Tutorials List</h4>
             <ul class="list-group">
@@ -34,9 +47,9 @@
                 <div>
                     <label><strong>Status:</strong></label> {{ currentTutorial.published ? "Published" : "Pending" }}
                 </div>
-                <a class="badge badge-warning" :href="'/tutorials/' + currentTutorial._id">
-                    Edit
-                </a>
+                <router-link :to="{ path: '/tutorials/' + currentTutorial._id }">
+                    <button type="button" class="btn btn-info">Edit</button>
+                </router-link>
             </div>
             <div v-else>
                 <br />
@@ -48,7 +61,7 @@
 
 <script>
 import TutorialDataService from "@/services/TutorialDataService";
-
+import Swal from 'sweetalert2'
 export default {
     name: 'tutorial-list',
     data() {
@@ -56,18 +69,58 @@ export default {
             tutorials: [],
             currentTutorial: null,
             currentIndex: -1,
-            title: ""
+            title: "",
+            page: 1,
+            count: 0,
+            pageSize: 5,
+            pageS: [3, 6, 9],
         }
     },
+    mounted() {
+        this.retriveTutorials();
+    },
     methods: {
+        getRequestParams(title, page, pageSize) {
+            let params = {};
+            if (title) {
+                params['title'] = title
+            }
+            if (page) {
+                params['page'] = page - 1;
+            }
+
+            if (pageSize) {
+                params['size'] = pageSize;
+            }
+
+            return params
+        },
+
         retriveTutorials() {
-            TutorialDataService.getAll()
+            const params = this.getRequestParams(
+                this.title,
+                this.page,
+                this.pageSize
+            )
+            TutorialDataService.getAll(params)
                 .then((res) => {
-                    this.tutorials = res.data.data
+                    const { data, totalItems } = res.data;
+                    this.tutorials = data;
+                    this.count = totalItems;
                 }).catch((error) => {
                     console.log(error);
                 })
         },
+        handlePageChange(value) {
+            this.page = value;
+            this.retriveTutorials()
+        },
+        handlePageSizeChange(event) {
+            this.pageSize = event.target.value;
+            this.page = 1;
+            this.retriveTutorials()
+        },
+
         refreshList() {
             this.retriveTutorials()
             this.currentTutorial = null;
@@ -78,13 +131,48 @@ export default {
             this.currentIndex = index
         },
         removeAllTutorials() {
-            TutorialDataService.deleteAll()
-                .then((res) => {
-                    this.refreshList();
-                }).catch((error) => {
-                    console.log(error);
+            const swalWithBootstrapButtons = Swal.mixin({
+                customClass: {
+                    confirmButton: 'btn btn-success',
+                    cancelButton: 'btn btn-danger'
+                },
+                buttonsStyling: false
+            })
+
+            swalWithBootstrapButtons.fire({
+                title: 'Are you sure?',
+                text: "You won't be able to revert this!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonText: 'Yes, delete it!',
+                cancelButtonText: 'No, cancel!',
+                reverseButtons: true
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    TutorialDataService.deleteAll()
+                        .then((res) => {
+                            this.refreshList();
+                        }).catch((error) => {
+                            console.log(error);
+                        }
+                        )
+                    swalWithBootstrapButtons.fire(
+                        'Deleted!',
+                        'Your file has been deleted.',
+                        'success'
+                    )
+
+                } else if (
+                    /* Read more about handling dismissals below */
+                    result.dismiss === Swal.DismissReason.cancel
+                ) {
+                    swalWithBootstrapButtons.fire(
+                        'Cancelled',
+                        'Your imaginary file is safe :)',
+                        'error'
+                    )
                 }
-                )
+            })
         },
         searchTitle() {
             TutorialDataService.findByTitle(this.title)
@@ -95,12 +183,7 @@ export default {
                 })
         }
 
-    },
-    mounted() {
-        this.retriveTutorials()
     }
-
-
 }
 
 </script>
